+++ src/main/java/it/vitalegi/archi/model/Container.java
@@ -2,8 +2,10 @@ package it.vitalegi.archi.model;
 
 import lombok.Data;
 import lombok.EqualsAndHashCode;
+import lombok.ToString;
 
 @EqualsAndHashCode(callSuper = true)
+@ToString(callSuper = true)
 @Data
 public class Container extends Element {
