+++ src/main/java/it/vitalegi/archi/model/Person.java
@@ -1,5 +1,9 @@
 package it.vitalegi.archi.model;
 
+import lombok.Data;
+import lombok.ToString;
+
+@ToString(callSuper = true)
 public class Person extends Element {
 
     public Element findChildById(String id) {
