+++ src/main/java/it/vitalegi/archi/model/SoftwareSystem.java
@@ -3,12 +3,14 @@ package it.vitalegi.archi.model;
 import it.vitalegi.archi.util.WorkspaceUtil;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
+import lombok.ToString;
 import lombok.extern.slf4j.Slf4j;
 
 import java.util.ArrayList;
 import java.util.List;
 @Slf4j
 @EqualsAndHashCode(callSuper = true)
+@ToString(callSuper = true)
 @Data
 public class SoftwareSystem extends Element {
     List<Container> containers;
